Move full text sources, data, citations/references dropdown out of Item component
